+++ src/location.rs
@@ -3,14 +3,14 @@ pub type InternedString = usize;
 #[derive(Debug, Clone, Copy)]
 pub struct Location {
     pub file: InternedString,
-    pub line: usize,
+    pub line: u32,
     pub span: Span,
 }
 
 #[derive(Debug, Clone, Copy)]
 pub struct Span {
-    pub start: usize,
-    pub len: usize,
+    pub start: u32,
+    pub len: u32,
 }
 
 impl Default for Location {
